Deployment/requirements.txt
streamlit
numpy
scikit-learn==1.0.2
tensorflow-cpu
scikit-image
pillow